Refactor git_config tests.
